+++ src/py42/services/detectionlists/high_risk_employee.py
@@ -1,4 +1,6 @@
 from py42.exceptions import Py42BadRequestError
+from py42.exceptions import Py42NotFoundError
+from py42.exceptions import Py42UserNotOnListError
 from py42.services import BaseService
 from py42.services.detectionlists import _DetectionListFilters
 from py42.services.detectionlists import _PAGE_SIZE
@@ -18,9 +20,7 @@ class HighRiskEmployeeFilters(_DetectionListFilters):
 class HighRiskEmployeeService(BaseService):
     """A service for interacting with High Risk Employee APIs."""
 
-    _api_version = u"v2"
-    _uri_prefix = u"/svc/api/{}".format(_api_version)
-    _resource = u"/highriskemployee"
+    _resource = u"v2/highriskemployee"
 
     def __init__(self, connection, user_context, user_profile_service):
         super(HighRiskEmployeeService, self).__init__(connection)
@@ -28,7 +28,7 @@ class HighRiskEmployeeService(BaseService):
         self._user_profile_service = user_profile_service
 
     def _make_uri(self, action):
-        return u"{}{}{}".format(self._uri_prefix, self._resource, action)
+        return u"{}{}".format(self._resource, action)
 
     def _add_high_risk_employee(self, tenant_id, user_id):
         data = {u"tenantId": tenant_id, u"userId": user_id}
@@ -90,7 +90,14 @@ class HighRiskEmployeeService(BaseService):
             u"userId": user_id,
         }
         uri = self._make_uri(u"/remove")
-        return self._connection.post(uri, json=data)
+        try:
+            return self._connection.post(uri, json=data)
+        except Py42NotFoundError as err:
+            message = (
+                "User with ID '{}' is not currently on the high-risk-employee "
+                "detection list.".format(user_id)
+            )
+            raise Py42UserNotOnListError(err, message=message)
 
     def get(self, user_id):
         """Gets user information.