+++ tests/conftest.py
@@ -6,6 +6,7 @@ from requests import HTTPError
 from requests import Response
 from requests import Session
 
+from py42.exceptions import Py42NotFoundError
 from py42.exceptions import Py42UnauthorizedError
 from py42.response import Py42Response
 from py42.sdk.queries.query_filter import QueryFilter
@@ -206,3 +207,13 @@ def create_mock_response(mocker, text):
     response.status_code = 200
     response.encoding = None
     return Py42Response(response)
+
+
+@pytest.fixture
+def mock_post_not_found_session(mocker, mock_connection):
+    response = mocker.MagicMock(spec=Response)
+    response.status_code = 404
+    exception = mocker.MagicMock(spec=HTTPError)
+    exception.response = response
+    mock_connection.post.side_effect = Py42NotFoundError(exception)
+    return mock_connection