+++ tests/services/detectionlists/test_departing_employee.py
@@ -8,6 +8,7 @@ from tests.conftest import TENANT_ID_FROM_RESPONSE
 
 from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42UserAlreadyAddedError
+from py42.exceptions import Py42UserNotOnListError
 from py42.services.detectionlists.departing_employee import DepartingEmployeeFilters
 from py42.services.detectionlists.departing_employee import DepartingEmployeeService
 from py42.services.detectionlists.user_profile import DetectionListUserService
@@ -136,9 +137,7 @@ class TestDepartingEmployeeClient(object):
             and posted_data["tenantId"] == _TENANT_ID_PARAM
             and posted_data["departureDate"] == "2022-12-20"
         )
-        assert (
-            mock_connection.post.call_args[0][0] == "/svc/api/v2/departingemployee/add"
-        )
+        assert mock_connection.post.call_args[0][0] == "v2/departingemployee/add"
         assert mock_connection.post.call_count == 2
 
     def test_add_when_user_already_on_list_raises_user_already_added_error(
@@ -181,10 +180,7 @@ class TestDepartingEmployeeClient(object):
             posted_data["userId"] == "999"
             and posted_data["tenantId"] == TENANT_ID_FROM_RESPONSE
         )
-        assert (
-            mock_connection.post.call_args[0][0]
-            == "/svc/api/v2/departingemployee/remove"
-        )
+        assert mock_connection.post.call_args[0][0] == "v2/departingemployee/remove"
 
     def test_get_all_posts_expected_data_to_expected_url(
         self,
@@ -209,10 +205,7 @@ class TestDepartingEmployeeClient(object):
             and posted_data["srtKey"] == "CREATED_AT"
             and posted_data["srtDirection"] == "DESC"
         )
-        assert (
-            mock_connection.post.call_args[0][0]
-            == "/svc/api/v2/departingemployee/search"
-        )
+        assert mock_connection.post.call_args[0][0] == "v2/departingemployee/search"
         assert mock_connection.post.call_count == 1
 
     def test_get_page_posts_data_to_expected_url(
@@ -243,10 +236,7 @@ class TestDepartingEmployeeClient(object):
             and posted_data["srtKey"] == "CREATED_AT"
             and posted_data["srtDirection"] == "DESC"
         )
-        assert (
-            mock_connection.post.call_args[0][0]
-            == "/svc/api/v2/departingemployee/search"
-        )
+        assert mock_connection.post.call_args[0][0] == "v2/departingemployee/search"
         assert mock_connection.post.call_count == 1
 
     def test_get_all_posts_expected_data_with_non_default_values(
@@ -266,10 +256,7 @@ class TestDepartingEmployeeClient(object):
 
         posted_data = mock_connection.post.call_args[1]["json"]
         assert mock_connection.post.call_count == 1
-        assert (
-            mock_connection.post.call_args[0][0]
-            == "/svc/api/v2/departingemployee/search"
-        )
+        assert mock_connection.post.call_args[0][0] == "v2/departingemployee/search"
         assert (
             posted_data["tenantId"] == user_context.get_current_tenant_id()
             and posted_data["filterType"] == "NEW_FILTER"
@@ -311,8 +298,7 @@ class TestDepartingEmployeeClient(object):
         mock_connection.post.return_value = mock_get_all_cases_response_empty
         client.set_alerts_enabled()
         assert (
-            mock_connection.post.call_args[0][0]
-            == "/svc/api/v2/departingemployee/setalertstate"
+            mock_connection.post.call_args[0][0] == "v2/departingemployee/setalertstate"
         )
 
     def test_get_posts_expected_data(
@@ -346,9 +332,7 @@ class TestDepartingEmployeeClient(object):
         )
         mock_connection.post.return_value = mock_get_all_cases_response_empty
         client.get("999")
-        assert (
-            mock_connection.post.call_args[0][0] == "/svc/api/v2/departingemployee/get"
-        )
+        assert mock_connection.post.call_args[0][0] == "v2/departingemployee/get"
 
     def test_update_posts_expected_data(
         self,
@@ -379,10 +363,7 @@ class TestDepartingEmployeeClient(object):
             mock_connection, user_context, mock_detection_list_user_client
         )
         client.update_departure_date(_USER_ID, "2022-12-20")
-        assert (
-            mock_connection.post.call_args[0][0]
-            == "/svc/api/v2/departingemployee/update"
-        )
+        assert mock_connection.post.call_args[0][0] == "v2/departingemployee/update"
 
     def test_update_posts_expected_data_with_datetime_instance(
         self,
@@ -405,4 +386,22 @@ class TestDepartingEmployeeClient(object):
             posted_data["userId"] == _USER_ID
             and posted_data["tenantId"] == TENANT_ID_FROM_RESPONSE
             and posted_data["departureDate"] == "2020-12-20"
+        )
+
+    def test_remove_raises_error_when_user_id_does_not_exist(
+        self,
+        user_context,
+        mock_post_not_found_session,
+        mock_detection_list_user_client,
+    ):
+        departing_employee_client = DepartingEmployeeService(
+            mock_post_not_found_session, user_context, mock_detection_list_user_client
+        )
+        user_id = "942897397520289999"
+        with pytest.raises(Py42UserNotOnListError) as err:
+            departing_employee_client.remove(user_id)
+        assert "User with ID '{}' is not currently on the departing-employee detection list.".format(
+            user_id
+        ) in str(
+            err.value
         )