+++ tests/test_exceptions.py
@@ -1,4 +1,5 @@
 import pytest
+from tests.conftest import REQUEST_EXCEPTION_MESSAGE
 
 from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42ForbiddenError
@@ -6,6 +7,7 @@ from py42.exceptions import Py42HTTPError
 from py42.exceptions import Py42InternalServerError
 from py42.exceptions import Py42MFARequiredError
 from py42.exceptions import Py42NotFoundError
+from py42.exceptions import Py42TooManyRequestsError
 from py42.exceptions import Py42UnauthorizedError
 from py42.exceptions import raise_py42_error
 
@@ -18,7 +20,7 @@ class TestPy42Errors(object):
 
     def test_raise_py42_error_raises_unauthorized_error(self, error_response):
         error_response.response.status_code = 401
-        with pytest.raises(Py42UnauthorizedError):
+        with pytest.raises(Py42UnauthorizedError, match=REQUEST_EXCEPTION_MESSAGE):
             raise_py42_error(error_response)
 
     def test_raise_py42_error_raises_MFA_required_error(self, error_response):
@@ -26,6 +28,10 @@ class TestPy42Errors(object):
         error_response.response.text = (
             '{"error":[{"primaryErrorKey":"TIME_BASED_ONE_TIME_PASSWORD_REQUIRED"}]}'
         )
+        with pytest.raises(Py42MFARequiredError):
+            raise_py42_error(error_response)
+
+        error_response.response.text = '{"error":[{"primaryErrorKey":"TOTP_AUTH_CONFIGURATION_REQUIRED_FOR_USER"}]}'
         with pytest.raises(Py42MFARequiredError):
             raise_py42_error(error_response)
 
@@ -49,7 +55,12 @@ class TestPy42Errors(object):
         with pytest.raises(Py42HTTPError):
             raise_py42_error(error_response)
 
-    @pytest.mark.parametrize("status_code", [400, 401, 403, 404, 500, 600])
+    def test_raise_py42_error_raises_too_many_requests_error(self, error_response):
+        error_response.response.status_code = 429
+        with pytest.raises(Py42TooManyRequestsError):
+            raise_py42_error(error_response)
+
+    @pytest.mark.parametrize("status_code", [400, 401, 403, 404, 429, 500, 600])
     def test_raise_py42_http_error_has_correct_response_type(
         self, error_response, status_code
     ):